# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  backend_internal_adapter_presenter.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
      total:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    name: Amplified Access
  description: Amplified Access Watchtower — incident reporting and monitoring platform
    API.
  title: Watchtower API
  version: "1.0"
paths:
...
  /admin/reports:
    get:
      description: Returns all reports belonging to the authenticated user's organization
      parameters:
      - description: Filter by status (draft, published, archived)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
      security:
      - BearerAuth: []
      summary: List organization reports
      tags:
      - Admin – Reports
    post:
      consumes:
      - application/json
      description: Creates a new report for the authenticated admin's organization
      parameters:
      - description: Report payload
        in: body
        name: body
        required: true
        schema:
          properties:
            fileKey:
              type: string
            status:
              type: string
            title:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
      security:
      - BearerAuth: []
      summary: Create report
      tags:
      - Admin – Reports
  /admin/reports/{id}:
    delete:
      description: Permanently removes a report
      parameters:
      - description: Report UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
      security:
      - BearerAuth: []
      summary: Delete report
      tags:
      - Admin – Reports
    get:
      description: Returns a single report by its UUID
      parameters:
      - description: Report UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
      security:
      - BearerAuth: []
      summary: Get report by ID
      tags:
      - Admin – Reports
    patch:
      consumes:
      - application/json
      description: Updates the title or status of an existing report
      parameters:
      - description: Report UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          properties:
            status:
              type: string
            title:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
      security:
      - BearerAuth: []
      summary: Update report
      tags:
      - Admin – Reports
  /admin/watchers:
    get:
      description: Returns all users with the watcher role. Accessible to admins and
        super admins.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
      security:
      - BearerAuth: []
      summary: List watchers
      tags:
      - Users
...
  /me:
    get:
      description: Returns the authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - Users
...
  /reports:
    get:
      description: Returns a paginated list of publicly visible reports
      parameters:
      - description: Page size (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Pagination offset
        in: query
        name: offset
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      - description: 'Sort field (default: created_at)'
        in: query
        name: sort
        type: string
      - description: 'Sort direction: asc or desc'
        in: query
        name: sortOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
      summary: List public reports
      tags:
      - Reports
  /reports/{id}:
    get:
      description: Returns a single published report by its UUID
      parameters:
      - description: Report UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
      summary: Get public report by ID
      tags:
      - Reports
...
  /superadmin/reports:
    get:
      description: Returns a paginated list of all reports across the platform (super
        admin only)
      parameters:
      - description: Page size (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Pagination offset
        in: query
        name: offset
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: 'Sort direction: asc or desc'
        in: query
        name: sortOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
      security:
      - BearerAuth: []
      summary: List all reports
      tags:
      - Super Admin – Reports
  /superadmin/users/admins:
    get:
      description: Returns all users with the admin role (super admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_adapter_presenter.Response'
      security:
      - BearerAuth: []
      summary: List admins
      tags:
      - Super Admin – Users
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token obtained from authentication. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"